Improve main/partyobj matching (gpmCol unsigned loop counter)
gpmCol trail-scan loop: cast counter to unsigned in the while test so
the bound check emits cmplwi 0x5 (matching target) not cmpwi.
gpmCol 75.79->75.88, unit 90.01818->90.01942.

+++ src/partyobj.cpp
@@ -4804,7 +4804,7 @@ void CGPartyObj::gpmCol()
 			newIndex = (i > trailIndex) ? trailIndex : i;
 			break;
 		}
-	} while (i < 5);
+	} while (static_cast<unsigned int>(i) < 5);
 
 	trailIndex = newIndex;
 #define gpmColClamp ((activeTrailCount - 1) & ~((activeTrailCount - 1) >> 31))
